Add support for DateTimeImmutable in buildPropertyBag
If `dtEnd` were to be a DateTimeImmutable instead of a DateTime, the add method would have no effect on the outcome.

Here, we update the stored dtEnd with the result of the add method.  For a DateTime object, this will be the same object - for a DateTimeImmutable object, this will be a new DateTimeImmutable one day in the future.