Fix calling TimeZone::createFromPhpDateTimeZone with default values
The default value for the begin date created a date B.C. which could not be parsed.
Therefore the new default value was set to the year of Christ birth.
The daylight saving time was introduced in 1916, therefore any begin date before that year will not produce further output.

Calling TimeZone::createFromPhpDateTimeZone with default values will now trigger a (silenced) deprecation error.

Fixes #240