Update EmailAddress.php
Several email/calendar clients do not support %40 in place of % in email addresses. For example, EmClient will show "nobody@invalid.invalid" instead of the actual email address. Also some other clients may work but show the %40 in plain text instead of converting it to @. 

So let's just not replace @ with %40. 😊